pip-requirements.txt
dbus-python
pyinotify
python-xlib
PyGObject
PyQt5
QScintilla
pyhamcrest
